little cleanup
