+++ zndraw/server/events.py
@@ -16,6 +16,7 @@ from zndraw.selection import Selection
 from zndraw.tasks import (
     load_zntrack_figures,
     load_zntrack_frames,
+    inspect_zntrack_node,
     run_analysis,
     run_geometry,
     run_modifier,
@@ -404,64 +405,29 @@ def init_socketio_events(io: SocketIO):
     @io.on("room:token:get")
     def room_token_get() -> str:
         return session.get("token")
+    
+    @io.on("zntrack:available")
+    def check_zntrack_available() -> bool:
+        try:
+            import zntrack
+            return True
+        except ImportError:
+            return False
+
 
     @io.on("zntrack:list-stages")
     def zntrack_list_stages(data: dict):
-        # room = session.get("token")
-        # emit("remote-file:fetch:enqueue", to=room)
-        # run_remote_file_fetch.delay(room, data)
-        # return ["NodeA", "NodeB", "NodeC"]
-        import dvc.api
+        try:
+            import dvc.api
 
-        fs = dvc.api.DVCFileSystem(url=data.get("remote"), rev=data.get("rev"))
-        return [x.name for x in fs.repo.stage.collect() if hasattr(x, "name")]
+            fs = dvc.api.DVCFileSystem(url=data.get("remote"), rev=data.get("rev"))
+            return [x.name for x in fs.repo.stage.collect() if hasattr(x, "name")]
+        except Exception as e:
+            return []
 
     @io.on("zntrack:inspect-stage")
     def zntrack_inspect_stage(data: dict):
-        import dataclasses
-
-        import zntrack
-        from zntrack.config import ZNTRACK_OPTION, ZnTrackOptionEnum
-
-        node = zntrack.from_rev(
-            name=data.get("name"), rev=data.get("rev"), remote=data.get("remote")
-        )
-
-        # find all @property decorated methods
-        def find_properties(cls):
-            exclude = set(
-                [
-                    "_init_descriptors_",
-                    "_init_subclass_basecls_",
-                    "_use_repr_",
-                    "nwd",
-                    "state",
-                    "uuid",
-                ]
-            )
-            names = []
-            for name in dir(cls):
-                if name not in exclude:
-                    attr = getattr(cls, name)
-                    if isinstance(attr, property):
-                        # get the type hint of the property
-                        hint = attr.fget.__annotations__.get("return", None)
-                        names.append((name, str(hint)))
-            return names
-
-        def find_deps_outs(cls):
-            names = []
-            for field in dataclasses.fields(cls):
-                if field.metadata.get(ZNTRACK_OPTION) in [
-                    ZnTrackOptionEnum.OUTS,
-                    ZnTrackOptionEnum.DEPS,
-                ]:
-                    # names.append(field.name)
-                    # append a tuple with the name and the type
-                    names.append((field.name, str(field.type)))
-            return names
-
-        return find_deps_outs(node) + find_properties(node.__class__)
+        return inspect_zntrack_node(**data)
 
     @io.on("zntrack:load-frames")
     def zntrack_load_frames(data: dict):
@@ -470,4 +436,4 @@ def init_socketio_events(io: SocketIO):
 
     @io.on("zntrack:load-figure")
     def zntrack_load_figure(data: dict):
-        load_zntrack_figures.delay(room=session.get("token"))
+        load_zntrack_figures.delay(room=session.get("token"), **data)
